refactor: integrate modular permission system with dual-mode support
Implement phase 0 of permission migration plan by creating seamless
integration layer between old and new permission systems. System can
now switch between implementations via USE_NEW_PERMISSION_SYSTEM env var.

Changes:
- Update all 15 API endpoints to use integration module imports
- Add adaptive functions that route to appropriate permission system
- Fix Pydantic v2 compatibility with PrivateAttr in Principal class
- Resolve circular imports between auth and permission modules
- Add missing Principal and get_current_permissions imports

Benefits:
- Zero-downtime migration capability
- Full backward compatibility maintained
- Instant rollback via environment variable
- No breaking changes to existing code

The system defaults to old permissions (USE_NEW_PERMISSION_SYSTEM=false)
ensuring stability while allowing gradual migration testing.
