+++ src/ctutor_backend/api/organizations.py
@@ -5,11 +5,13 @@ from gitlab import Gitlab
 from pydantic import BaseModel
 from sqlalchemy.orm import Session
 from ctutor_backend.api.auth import get_current_permissions
+from ctutor_backend.permissions.integration import (
+    adaptive_check_permissions as check_permissions,
+    Principal
+)
 from ctutor_backend.api.exceptions import BadRequestException, NotImplementedException
-from ctutor_backend.api.permissions import check_permissions
 from ctutor_backend.database import get_db
 from ctutor_backend.interface.organizations import OrganizationInterface, OrganizationProperties
-from ctutor_backend.interface.permissions import Principal
 from ctutor_backend.api.api_builder import CrudRouter
 from ctutor_backend.interface.tokens import encrypt_api_key
 from ctutor_backend.model.organization import Organization
