+++ src/ctutor_backend/api/results.py
@@ -2,10 +2,10 @@ from typing import Annotated
 from uuid import UUID
 from fastapi import Depends
 from ctutor_backend.api.api_builder import CrudRouter
+
 from ctutor_backend.api.auth import get_current_permissions
-from ctutor_backend.api.permissions import check_permissions
+from ctutor_backend.permissions.integration import adaptive_check_permissions as check_permissions, Principal
 from ctutor_backend.database import get_db
-from ctutor_backend.interface.permissions import Principal
 from ctutor_backend.interface.results import ResultInterface, ResultStatus
 from ctutor_backend.model.result import Result
 from ctutor_backend.tasks import get_task_executor
