+++ src/ctutor_backend/api/role_claims.py
@@ -1,14 +1,12 @@
 from typing import Annotated
-from fastapi import Depends
-from fastapi import APIRouter
+from fastapi import Depends, APIRouter
 from sqlalchemy.orm import Session
+
+from ctutor_backend.permissions.integration import adaptive_check_permissions as check_permissions, Principal
 from ctutor_backend.api.auth import get_current_permissions
-from ctutor_backend.api.permissions import check_permissions
 from ctutor_backend.database import get_db
-from ctutor_backend.interface.permissions import Principal
 from ctutor_backend.interface.roles_claims import RoleClaimList, RoleClaimQuery, role_claim_search
 from ctutor_backend.model.role import RoleClaim
-
 role_claim_router = APIRouter()
 
 @role_claim_router.get("", response_model=list[RoleClaimList])
