+++ src/ctutor_backend/api/user_roles.py
@@ -5,15 +5,14 @@ from fastapi import APIRouter
 from sqlalchemy.orm import Session
 from sqlalchemy import exc
 from ctutor_backend.api.exceptions import InternalServerException
-from ctutor_backend.api.auth import get_current_permissions
+
 from ctutor_backend.api.crud import create_db, list_db
 from ctutor_backend.api.exceptions import NotFoundException
-from ctutor_backend.api.permissions import check_permissions
+from ctutor_backend.api.auth import get_current_permissions
+from ctutor_backend.permissions.integration import adaptive_check_permissions as check_permissions, Principal
 from ctutor_backend.database import get_db
-from ctutor_backend.interface.permissions import Principal
 from ctutor_backend.interface.user_roles import UserRoleCreate, UserRoleGet, UserRoleInterface, UserRoleList, UserRoleQuery
 from ctutor_backend.model.role import UserRole
-
 user_roles_router = APIRouter()
 
 @user_roles_router.get("", response_model=list[UserRoleList])
